// src/components/jtl/table/types/ColumnSizeInput.ts
/**
 * Column size properties used for computing width-related styles
 */
type ColumnSizeInput = {
  /** Column width in pixels */
  width?: number;
  /** Minimum column width in pixels */
  minWidth?: number;
  /** Maximum column width in pixels */
  maxWidth?: number;
};

export default ColumnSizeInput;